| 9510    | Support ${HOST} substitution in "From" field. Thanks to      |
   | zmedico | wschlich.                                                    |
   
   | 9511    | Move the default empty compiler flag settings from ebuild.sh |
   | zmedico | to make.globals (bug #214499).                               |
   
   | 9512    | Bug #214879 - In fetch(), only parse the Manifest when       |
   | zmedico | FEATURES=strict is enabled.                                  |
   
   | 9513    | Make --noreplace take precedence over --newuse, as suggested |
   | zmedico | by wolf31o2.                                                 |
   
   | 9514    | Fix the noreplace logic to break out of the cpv loop early   |
   | zmedico | since there's no point in examining any lower versions from  |
   |         | the same repo.                                               |
   
   | 9519    | Bug #214879 - Fix FEATURES=-strict behavior so that          |
   | zmedico | distfiles digests are always checked.                        |
   
   | 9520    | Bug #212509 - Implement a new --without-mask option that     |
   | zmedico | causes repoman to behave as if there are no package.mask     |
   |         | entries.                                                     |
   
   | 9521    | Bug #212509 - Display a note about the --without-mask option |
   | zmedico | when there are masked packages.                              |
   
   | 9522    | Bug #209418 - Add a new KEYWORDS.dropped check that compares |
   | zmedico | keywords against the keywords of earlier ebuild versions     |
   |         | within a slot.                                               |
   
   | 9523    | Remove redundant info from the KEYWORDS.dropped message.     |
   | zmedico |                                                              |
   
   | 9524    | Bug #213629 - Create an EAPI.incompatible category and use   |
   | zmedico | it in cases where EAPI=0 and a slot atom is encountered.     |
   
   | 9525    | Bug #213629 - Use EAPI.incompatible in cases where EAPI=0    |
   | zmedico | and a default IUSE is encountered.                           |
   
   | 9526    | Bug #213698 - Make FEAUTURES=keepwork and keeptemp preserve  |
   | zmedico | ${T}/environment since it may contain essential state.       |
   
   | 9527    | Make dyn_clean() consistent wrt removal of                   |
   | zmedico | ${PORTAGE_BUILDDIR}/image and ${PORTAGE_BUILDDIR}/.installed |
   |         | since otherwise with FEATURES=keepwork dyn_install() will    |
   |         | not call src_install() as necessary to repopulate            |
   |         | ${PORTAGE_BUILDDIR}/image (resulting in install_qa_check     |
   |         | failure due to absence of ${PORTAGE_BUILDDIR}/image).        |
   
   | 9528    | Bug #214619 - Add support for a PORTAGE_BINPKG_TAR_OPTS      |
   | zmedico | variable that allows the user to specify tar command options |
   |         | for binary package creation.                                 |
   
   | 9537    | Disable tarsync when using lzma compression since it doesn't |
   | zmedico | seem to be supported.                                        |
   
   | 9538    | * Skip the recusive chown call if it fails on the top level  |
   | zmedico | directory since it's possible that emerge-webrsync is not    |
   |         | run by root. * Just chown to portage:portage instead of      |
   |         | using PORTAGE_INST_{UID,GID}.  Prefix: use PORTAGE_USER and  |
   |         | PORTAGE_GROUP instead.                                       |

svn path=/main/branches/prefix/; revision=9583

+++ bin/ebuild.sh
@@ -273,12 +273,6 @@ export LIBOPTIONS="-m0644"
 export DIROPTIONS="-m0755"
 export MOPREFIX=${PN}
 
-# When compiler flags are unset, many packages will substitute their own
-# implicit flags. For uniformity, use an empty string as the default.
-[ "${CFLAGS-unset}"    == "unset" ] && export CFLAGS=""
-[ "${CXXFLAGS-unset}"  == "unset" ] && export CXXFLAGS=""
-[ "${LDFLAGS-unset}"   == "unset" ] && export LDFLAGS=""
-
 check_KV() {
 	if [ -z "${KV}" ]; then
 		eerror ""
@@ -676,11 +670,9 @@ dyn_unpack() {
 	if [ "${newstuff}" == "yes" ]; then
 		# We don't necessarily have privileges to do a full dyn_clean here.
 		rm -rf "${WORKDIR}"
-		if [ -d "${T}" ] && ! hasq keeptemp ${FEATURES} ; then
+		if [ -d "${T}" ] && \
+			! hasq keeptemp $FEATURES && ! hasq keepwork $FEATURES ; then
 			rm -rf "${T}" && mkdir "${T}"
-		else
-			[ -e "${T}/environment" ] && \
-				mv "${T}/environment" "${T}/environment.keeptemp"
 		fi
 	fi
 	if [ -e "${WORKDIR}" ]; then
@@ -717,11 +709,10 @@ dyn_clean() {
 	fi
 
 	rm -rf "${PORTAGE_BUILDDIR}/image" "${PORTAGE_BUILDDIR}/homedir"
+	rm -f "${PORTAGE_BUILDDIR}/.installed"
 
-	if ! hasq keeptemp $FEATURES; then
+	if ! hasq keeptemp $FEATURES && ! hasq keepwork $FEATURES ; then
 		rm -rf "${T}"
-	else
-		[ -e "${T}/environment" ] && mv "${T}/environment" "${T}/environment.keeptemp"
 	fi
 
 	if ! hasq keepwork $FEATURES; then
@@ -730,7 +721,6 @@ dyn_clean() {
 		rm -rf "${PORTAGE_BUILDDIR}/.unpacked"
 		rm -rf "${PORTAGE_BUILDDIR}/.compiled"
 		rm -rf "${PORTAGE_BUILDDIR}/.tested"
-		rm -rf "${PORTAGE_BUILDDIR}/.installed"
 		rm -rf "${PORTAGE_BUILDDIR}/.packaged"
 		rm -rf "${PORTAGE_BUILDDIR}/build-info"
 		rm -rf "${WORKDIR}"
